#266 basic_cat examples not building on MSVC
Replacing the home-grown tuple_push_back function with std::tuple_cat causes MSVC to hang.  I can't reproduce on CE so it must be related to some of the more complex types used in the basic_cat example.

I also set build timeouts in the GHAs because this issue causes a 6 hour hang...

Fix #266